MyRectangle completed and tested

+++ 211Project1/src/MyShape.java
@@ -2,16 +2,16 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
 public class MyShape {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private javafx.scene.paint.Color color;
 
-    MyShape(int x, int y){
+    MyShape(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -19,11 +19,11 @@ public class MyShape {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
